AER-2710 rename WNB to OWN2000
Keep Theme WNB still present to avoid breaking backward compatibility.
