+++ source/imaer-shared/src/main/java/nl/overheid/aerius/shared/domain/calculation/CalculationSetOptions.java
@@ -38,7 +38,7 @@ public class CalculationSetOptions implements Serializable {
   private boolean useInCombinationArchive;
   private final ArrayList<Substance> substances = new ArrayList<>();
   private final Set<EmissionResultKey> emissionResultKeys = new HashSet<>();
-  private WNBCalculationOptions wnbCalculationOptions = new WNBCalculationOptions();
+  private OwN2000CalculationOptions own2000CalculationOptions = new WNBCalculationOptions();
   private RBLCalculationOptions rblCalculationOptions = new RBLCalculationOptions();
   private NCACalculationOptions ncaCalculationOptions = new NCACalculationOptions();
   private Serializable experimentalOptions;
@@ -134,8 +134,16 @@ public class CalculationSetOptions implements Serializable {
     this.connectSuppliedOptions = connectSuppliedOptions;
   }
 
+  public OwN2000CalculationOptions getOwN2000CalculationOptions() {
+    return own2000CalculationOptions;
+  }
+
+  /**
+   * @deprecated replace calls with {@link #getOwN2000CalculationOptions()}
+   */
+  @Deprecated
   public WNBCalculationOptions getWnbCalculationOptions() {
-    return wnbCalculationOptions;
+    return (WNBCalculationOptions) own2000CalculationOptions;
   }
 
   public RBLCalculationOptions getRblCalculationOptions() {
@@ -164,7 +172,7 @@ public class CalculationSetOptions implements Serializable {
   public String toString() {
     return "CalculationSetOptions [calculationSetOptionsId=" + calculationSetOptionsId + ", calculationMethod=" + calculationMethod
         + ", calculateMaximumRange=" + calculateMaximumRange + ", substances=" + substances + ", emissionResultKeys=" + emissionResultKeys
-        + ", stacking=" + stacking + ", connectSuppliedOptions=" + connectSuppliedOptions + ", wnbCalculationOptions=" + wnbCalculationOptions
+        + ", stacking=" + stacking + ", connectSuppliedOptions=" + connectSuppliedOptions + ", own2000CalculationOptions=" + own2000CalculationOptions
         + ", rblCalculationOptions=" + rblCalculationOptions + ", ncaCalculationOptions=" + ncaCalculationOptions + "]";
   }
 }
